Remove the `Four` Metacategory
It isn’t used anywhere and it takes a long time to construct. Note that
the construction is done automatically by `Program` and the `Obligation
Tactic`, but since a lot of cases are considered, it takes a long time.

If ever `Four` ever becomes necessary or useful, here are two ways to
introduce it again:
- revert this commit, thus adding the long compilation time back again.
- prove a general theorem that each `{| pairs := composable_pairs n |}`
for `n : nat` defines a Metacategory.