Give explicit proofs in Construction.Comma
This reduces compilation time for this file by a factor of ~2/3 on my
machine. It does this, by not letting `cat_simpl` search a proof, but
giving it explicitly and trying to do as little work in the proof as
possible (destruct & unfold as little as possible).

Time measurements: From 64 seconds to 37 seconds, measured using `time
make Construction/Comma.vo`.